#10334: Migrate ComplexTensor struct to TTNN
