+++ ttnn/cpp/ttnn/operations/eltwise/binary/binary_pybind.hpp
@@ -12,6 +12,7 @@
 #include "ttnn/operations/eltwise/binary/binary_composite.hpp"
 #include "ttnn/operations/eltwise/complex_binary/complex_binary.hpp"
 #include "ttnn/types.hpp"
+#include "ttnn/operations/eltwise/complex_binary/device/complex_binary_op.hpp"
 
 namespace py = pybind11;
 
@@ -21,6 +22,8 @@ namespace operations {
 namespace binary {
 
 namespace detail {
+
+using ComplexTensor = complex_binary::ComplexTensor;
 
 template <typename binary_operation_t>
 void bind_binary_operation(py::module& module, const binary_operation_t& operation, const std::string& description) {
