+++ ttnn/cpp/ttnn/operations/eltwise/complex_binary/device/complex_binary_op.hpp
@@ -20,6 +20,34 @@ enum class ComplexBinaryOpType {
     DIV,
 };
 
+class ComplexTensor {
+    private:
+        std::array<Tensor,2> m_real_imag;
+
+    public:
+
+        ComplexTensor(std::array<Tensor,2> val): m_real_imag(val) {
+            TT_ASSERT( m_real_imag[0].get_legacy_shape() == m_real_imag[1].get_legacy_shape() , "Tensor shapes of real and imag should be identical");
+        }
+
+        const Tensor& operator[](uint32_t index) const {
+            return m_real_imag[index];
+        }
+
+        const Tensor& real() const {
+            return m_real_imag[0];
+        }
+
+        const Tensor& imag() const {
+            return m_real_imag[1];
+        }
+
+        void deallocate() {
+            m_real_imag[0].deallocate();
+            m_real_imag[1].deallocate();
+        }
+};
+
 // OpHandler_complex_binary_type1 = get_function_complex_binary
 ComplexTensor _add(const ComplexTensor& input_a, const ComplexTensor& input_b, const MemoryConfig& output_mem_config);
 ComplexTensor _sub(const ComplexTensor& input_a, const ComplexTensor& input_b, const MemoryConfig& output_mem_config);
