+++ ttnn/cpp/ttnn/operations/eltwise/complex_binary_backward/complex_binary_backward_pybind.hpp
@@ -19,6 +19,7 @@ namespace operations {
 namespace complex_binary_backward {
 
 namespace detail {
+using ComplexTensor = complex_binary::ComplexTensor;
 
 template <typename complex_binary_backward_operation_t>
 void bind_complex_binary_backward_w_float(py::module& module, const complex_binary_backward_operation_t& operation, const std::string& description) {
