+++ ttnn/cpp/ttnn/operations/eltwise/complex_binary_backward/device/complex_binary_backward_op.cpp
@@ -12,6 +12,7 @@
 
 
 namespace ttnn::operations::complex_binary_backward {
+using ComplexTensor = complex_binary::ComplexTensor;
 
 // complex add
 // self: grad, other: grad * alpha
